refactor: moved upgradeforms to another file
